feat: login

+++ src/main/java/com/example/dishpatch/infra/db/store/repository/StoreRepository.java
@@ -1,8 +1,7 @@
 package com.example.dishpatch.infra.db.store.repository;
 
+import com.example.dishpatch.infra.db.store.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.dishpatch.infra.db.store.entity.Store;
-
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+public interface StoreRepository extends JpaRepository<Store, Long> {
 }
